nbody-master/requirements.txt

PyOpenGL==3.1.0
PyQt5==5.8.1.1
pyrr==0.10.3
